college/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
cashfree-sdk==0.1.3
celery==5.3.5
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap4==2023.1
cryptography==41.0.4
Django==4.2.6
django-countries==7.5.1
django-crispy-forms==2.1
django-paypal==2.0
django-redis==5.4.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Faker==20.0.3
idna==3.4
kombu==5.3.4
mysql-connector-python==8.2.0
mysqlclient==2.1.0
paypalrestsdk==1.13.1
pem==23.1.0
Pillow==10.1.0
prompt-toolkit==3.0.41
protobuf==4.21.12
pycparser==2.21
pycryptodome==3.19.0
PyJWT==2.8.0
PyMySQL==1.1.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
razorpay==1.4.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stripe==7.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.11
